Move all vcpu/numa code to the vcpu module and misc refreshes
