+++ pkg/virt-launcher/virtwrap/converter/BUILD.bazel
@@ -6,9 +6,7 @@ go_library(
         "converter.go",
         "generated_mock_converter.go",
         "network.go",
-        "numa_placement.go",
         "pci-placement.go",
-        "vcpu_placement.go",
     ],
     importpath = "kubevirt.io/kubevirt/pkg/virt-launcher/virtwrap/converter",
     visibility = ["//visibility:public"],
@@ -34,7 +32,6 @@ go_library(
         "//vendor/github.com/golang/mock/gomock:go_default_library",
         "//vendor/golang.org/x/sys/unix:go_default_library",
         "//vendor/k8s.io/api/core/v1:go_default_library",
-        "//vendor/k8s.io/apimachinery/pkg/api/resource:go_default_library",
     ],
 )
 
@@ -43,7 +40,6 @@ go_test(
     srcs = [
         "converter_suite_test.go",
         "converter_test.go",
-        "numa_placement_test.go",
     ],
     embed = [":go_default_library"],
     deps = [
