+++ pkg/virt-launcher/virtwrap/converter/vcpu/numa_placement_test.go
@@ -1,4 +1,4 @@
-package converter
+package vcpu
 
 import (
 	"strconv"
